// lib/database.types.ts
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[]

export interface Database {
  public: {
    Tables: {
      profiles: {
        Row: {
          id: string
          email: string
          full_name: string | null
          avatar_url: string | null
          created_at: string
        }
        Insert: {
          id: string
          email: string
          full_name?: string | null
          avatar_url?: string | null
          created_at?: string
        }
        Update: {
          id?: string
          email?: string
          full_name?: string | null
          avatar_url?: string | null
          created_at?: string
        }
      }
      products: {
        Row: {
          id: string
          user_id: string
          name: string
          description: string | null
          price: number | null
          created_at: string
          updated_at: string
          template: string
          slug: string
          published: boolean
          product_type: string
        }
        Insert: {
          id?: string
          user_id: string
          name: string
          description?: string | null
          price?: number | null
          created_at?: string
          updated_at?: string
          template: string
          slug: string
          published?: boolean
          product_type: string
        }
        Update: {
          id?: string
          user_id?: string
          name?: string
          description?: string | null
          price?: number | null
          created_at?: string
          updated_at?: string
          template?: string
          slug?: string
          published?: boolean
          product_type?: string
        }
      }
    }
  }
}